@@ -6,14 +6,14 @@
     "@testing-library/jest-dom": "^5.14.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "firebase": "^9.9.4",
+    "firebase": "^9.13.0",
     "materialize-css": "^1.0.0",
-    "react": "^17.0.2",
-    "react-dnd": "^14.0.3",
-    "react-dnd-html5-backend": "^14.0.1",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
+    "react-dom": "^18.2.0",
     "react-firebase-hooks": "^3.0.4",
-    "react-router-dom": "^5.2.1",
+    "react-router-dom": "^6.4.2",
     "react-scripts": "^5.0.1",
     "web-vitals": "^1.1.2"
   },
